requirements.txt
torch
torchvision
opencv-python
numpy
pandas
tqdm
pyyaml
matplotlib
pillow
